+++ packages/android_intent/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## NEXT
 
 * Remove references to the V1 Android embedding.
+* Updated Android lint settings.
+* Specify Java 8 for Android build.
 
 ## 2.0.2