+++ packages/path_provider/path_provider/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.0.5
+
+* Update minimum Flutter SDK to 2.5 and iOS deployment target to 9.0.
+
+## 2.0.4
+
+* Updated Android lint settings.
+* Specify Java 8 for Android build.
+
 ## 2.0.3
 
 * Add iOS unit test target.