+++ packages/url_launcher/url_launcher/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 6.0.12
+
+* Fixed an error where 'launch' method of url_launcher would cause an error if the provided URL was not valid by RFC 3986.
+
+## 6.0.11
+
+* Update minimum Flutter SDK to 2.5 and iOS deployment target to 9.0.
+* Updated Android lint settings.
+
 ## 6.0.10
 
 * Remove references to the Android v1 embedding.